refactor: replace Datasets page with Notebook Page
